+++ dockci/workers.py
@@ -6,7 +6,7 @@ import os
 
 from flask_mail import Message
 
-from dockci.models.build import Build
+from dockci.models.job import Job
 from dockci.models.project import Project
 from dockci.server import APP, MAIL
 from dockci.notifications import HipChat
@@ -26,40 +26,40 @@ def send_mail(message):
             )
 
 
-def run_build_async(project_slug, build_slug):
+def run_job_async(project_slug, job_slug):
     """
-    Load and run a build's private run project, forking to handle the build in
+    Load and run a job's private run project, forking to handle the job in
     the background
     """
     if os.fork():
         return  # parent process
 
-    logger = logging.getLogger('dockci.build')
+    logger = logging.getLogger('dockci.job')
     try:
         with APP.app_context():
             project = Project(project_slug)
-            build = Build(project=project, slug=build_slug)
-            build_okay = build._run_now()  # pylint:disable=protected-access
+            job = Job(project=project, slug=job_slug)
+            job_okay = job._run_now()  # pylint:disable=protected-access
 
             # Send the failure message
-            if not build_okay:
+            if not job_okay:
                 recipients = []
-                if build.git_author_email:
+                if job.git_author_email:
                     recipients.append('%s <%s>' % (
-                        build.git_author_name,
-                        build.git_author_email
+                        job.git_author_name,
+                        job.git_author_email
                     ))
-                if build.git_committer_email:
+                if job.git_committer_email:
                     recipients.append('%s <%s>' % (
-                        build.git_committer_name,
-                        build.git_committer_email
+                        job.git_committer_name,
+                        job.git_committer_email
                     ))
 
                 if recipients:
                     subject = (
-                        "DockCI - {project_name} {build_result}ed".format(
+                        "DockCI - {project_name} {job_result}ed".format(
                             project_name=project.name,
-                            build_result=build.result,
+                            job_result=job.result,
                         )
                     )
                     email = Message(
@@ -72,11 +72,11 @@ def run_build_async(project_slug, build_slug):
             if project.hipchat_api_token != '' and project.hipchat_room != '':
                 hipchat = HipChat(apitoken=project.hipchat_api_token,
                                   room=project.hipchat_room)
-                hipchat.message("DockCI - {name} Build {id}: {result}".format(
+                hipchat.message("DockCI - {name} Job {id}: {result}".format(
                     name=project.name,
-                    id=build.create_ts,
-                    result=build.result,
+                    id=job.create_ts,
+                    result=job.result,
                 ))
 
     except Exception:  # pylint:disable=broad-except
-        logger.exception("Something went wrong in the build worker")
+        logger.exception("Something went wrong in the job worker")